Handle lil TTL
